+++ src/index.ts
@@ -1,21 +1,5 @@
 export { StructBuffer, sizeof, display } from "./struct-buffer";
-export {
-  BYTE,
-  WORD,
-  DWORD,
-  QWORD,
-  int8_t,
-  int16_t,
-  int32_t,
-  int64_t,
-  uint8_t,
-  uint16_t,
-  uint32_t,
-  uint64_t,
-  float,
-  double,
-  char,
-  string_t,
-} from "./types";
+export * from "./types";
 
-export { registerType } from "./class-type";
+export { registerType, typedef } from "./class-type";
+export { parseCStruct, defaultTypes } from "./parse-cstruct";